Correct spelling in package name
