+++ dev/com.ibm.ws.microprofile.faulttolerance.1.0.cdi.services/src/com/ibm/ws/microprofile/faulttolerance10/cdi/services/FTEnablementConfigComponent.java
@@ -10,13 +10,13 @@
  * Contributors:
  *     IBM Corporation - initial API and implementation
  *******************************************************************************/
-package com.ibm.ws.micrprofile.faulttolerance10.cdi.services;
+package com.ibm.ws.microprofile.faulttolerance10.cdi.services;
 
 import org.osgi.service.component.annotations.Component;
 import org.osgi.service.component.annotations.ConfigurationPolicy;
 
-import com.ibm.ws.microprofile.faulttolerance.cdi.config.AnnotationConfigFactory;
-import com.ibm.ws.microprofile.faulttolerance.cdi.config.impl.AnnotationConfigFactoryImpl;
+import com.ibm.ws.microprofile.faulttolerance.cdi.FTEnablementConfig;
+import com.ibm.ws.microprofile.faulttolerance.cdi.config.impl.FTEnablementConfig10Impl;
 
-@Component(service = AnnotationConfigFactory.class, configurationPolicy = ConfigurationPolicy.IGNORE)
-public class AnnotationConfigFactoryComponent extends AnnotationConfigFactoryImpl {}
+@Component(service = FTEnablementConfig.class, configurationPolicy = ConfigurationPolicy.IGNORE)
+public class FTEnablementConfigComponent extends FTEnablementConfig10Impl {}
