--- dev/com.ibm.ws.microprofile.faulttolerance.1.0.cdi.services/src/com/ibm/ws/micrprofile/faulttolerance10/cdi/services/package-info.java
@@ -1,19 +1,0 @@
-/*******************************************************************************
- * Copyright (c) 2018 IBM Corporation and others.
- * All rights reserved. This program and the accompanying materials
- * are made available under the terms of the Eclipse Public License 2.0
- * which accompanies this distribution, and is available at
- * http://www.eclipse.org/legal/epl-2.0/
- * 
- * SPDX-License-Identifier: EPL-2.0
- *
- * Contributors:
- *     IBM Corporation - initial API and implementation
- *******************************************************************************/
-/**
- * Components for Fault Tolerance 1.0
- */
-@TraceOptions(traceGroup = "FAULTTOLERANCE")
-package com.ibm.ws.micrprofile.faulttolerance10.cdi.services;
-
-import com.ibm.websphere.ras.annotation.TraceOptions;
