+++ dev/com.ibm.ws.microprofile.faulttolerance.1.1.cdi.services/src/com/ibm/ws/microprofile/faulttolerance11/cdi/services/AnnotationConfigFactoryComponent.java
@@ -10,7 +10,7 @@
  * Contributors:
  *     IBM Corporation - initial API and implementation
  *******************************************************************************/
-package com.ibm.ws.micrprofile.faulttolerance11.cdi.services;
+package com.ibm.ws.microprofile.faulttolerance11.cdi.services;
 
 import org.osgi.service.component.annotations.Component;
 import org.osgi.service.component.annotations.ConfigurationPolicy;
