+++ dev/com.ibm.ws.microprofile.faulttolerance.2.0.cdi.services/src/com/ibm/ws/microprofile/faulttolerance20/cdi/services/FTEnablementConfigComponent.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2019 IBM Corporation and others.
+ * Copyright (c) 2018 IBM Corporation and others.
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License 2.0
  * which accompanies this distribution, and is available at
@@ -10,7 +10,7 @@
  * Contributors:
  *     IBM Corporation - initial API and implementation
  *******************************************************************************/
-package com.ibm.ws.micrprofile.faulttolerance21.cdi.services;
+package com.ibm.ws.microprofile.faulttolerance20.cdi.services;
 
 import org.osgi.service.component.annotations.Component;
 import org.osgi.service.component.annotations.ConfigurationPolicy;
@@ -19,5 +19,4 @@ import com.ibm.ws.microprofile.faulttolerance.cdi.FTEnablementConfig;
 import com.ibm.ws.microprofile.faulttolerance.cdi.config.impl.FTEnablementConfig11Impl;
 
 @Component(service = FTEnablementConfig.class, configurationPolicy = ConfigurationPolicy.IGNORE)
-public class FTEnablementConfigComponent extends FTEnablementConfig11Impl {
-}
+public class FTEnablementConfigComponent extends FTEnablementConfig11Impl {}
