+++ dev/com.ibm.ws.microprofile.faulttolerance.2.1.cdi.services/src/com/ibm/ws/microprofile/faulttolerance21/cdi/services/FTEnablementConfigComponent.java
@@ -1,0 +1,23 @@
+/*******************************************************************************
+ * Copyright (c) 2019 IBM Corporation and others.
+ * All rights reserved. This program and the accompanying materials
+ * are made available under the terms of the Eclipse Public License 2.0
+ * which accompanies this distribution, and is available at
+ * http://www.eclipse.org/legal/epl-2.0/
+ * 
+ * SPDX-License-Identifier: EPL-2.0
+ *
+ * Contributors:
+ *     IBM Corporation - initial API and implementation
+ *******************************************************************************/
+package com.ibm.ws.microprofile.faulttolerance21.cdi.services;
+
+import org.osgi.service.component.annotations.Component;
+import org.osgi.service.component.annotations.ConfigurationPolicy;
+
+import com.ibm.ws.microprofile.faulttolerance.cdi.FTEnablementConfig;
+import com.ibm.ws.microprofile.faulttolerance.cdi.config.impl.FTEnablementConfig11Impl;
+
+@Component(service = FTEnablementConfig.class, configurationPolicy = ConfigurationPolicy.IGNORE)
+public class FTEnablementConfigComponent extends FTEnablementConfig11Impl {
+}
